Shorten backend package name to 'app'
This follows the naming convention used in FastAPI Project Template,
which is a project generator maintained by the same author of FastAPI.

https://github.com/tiangolo/full-stack-fastapi-postgresql

The structure of the base project generated is like this:
'base-project/': The top-level directory.
'backend/': The directory for backend stuff.
'app/': An extra layer for docker compose, which does not have a
counterpart in this project yet.
'app/': The python package for fastapi.

'app' is the default fastapi package name suggested.
And there was no config promt to change this while intializing the
project.
So I guess 'app' is a good name. The only problem, I can think of now,
using such a general or indistinguishable name would be when we need
another fastapi package.
Would reflect the naming again when that actually happens.