Revert update_static to sequential mode
There happens some unresolved bug when using asyncio to parallelize the
requests to LoL static API.

The async version is kept in the backend-dev-update-static-async.

Since the reimplemented sequetial mode seems to be working fast enough,
the async version will probably not be reworked in a near future.

Lastly, tqdm used to add a simple progress bar.