Add early stop url call
This url call is provided by Azure Logic App.
When this url is called, the Azure Logic App initiate the deallocating
the of the virtual machine instance.

The instance is also regularly stopped at 00:30:00(which would be enough
for update script to finish if it operated normally) by another Azure
Logic App, in case some bugs stopped the script from finishing.