Fix League Client connection issue
Automated price history update on KST20210823 didn't work.
I did an more thorough investigation of the cause the errors raised by
League Client driver.

THe analysis is written in the README in league-client-stuff.

Hopefully, it shouldn't bothered by the same problem again. Although it
is not a very elegant fix.