Refactor database configuration logic
Now, database selection logic is moved to 'db_config.py'.

'database.py' is now intended to be ignorant of the database in use.
This means it cannot set the 'check_same_thread' properly on connection
to SQLite, which is only required in case of SQLite.
However, since SQlite is only used for local testing, there wouldn't be
multiple connections that could probably raise error.
So it would cause no harm in my opinion.

'db_config.py' tries to load 'db_config.ini' in current directory(I
recommended it to be the 'backend' directory). If the config file is not
found or not properly set, an local SQLite would instead be referenced.

Note that 'db_config.py' is not tracked! So that it can survive git
pulls and only need to be configured once.

Instead, 'db_config_template.ini' is provided for easy start up.

'.gitignore' is changed accordingly. The original 'db_account' is
removed from untrack list, and the new 'db_config.ini' is added.