Replace DDL statement with alembic
All DDL statement(specifically the 'create_all()' from SQLAlchemy) is
successfully removed with usage of alembic.

Current, AWS RDS database is stamped as the latest revision, as well as
the local 'sqlite3.db'.

The local 'sqlite3.db' is, however, planned to be removed soon.
It's going to be replaced with in memory test database generated each
time on reference.