Change 'db_config.ini' file search directory
Now, 'db_config.ini' has to be located in backend directory regaredless
of the current directory, which is good!

Previously, if one run some script that load db_config in a folder other
backend, 'db_config.ini' in the backend would not be loaded.

For example, '/path/that/is/not/backend/$ python some_script.py' would
failed to load 'db_config.ini' in the backend folder.

But now, search for the 'db_config.ini' is relative to the location of
'db_config.py'. 'db_config.py' will always search the config file in
it's parent folder.