Revert "Add database trigger for 'price' and 'sale_price'"
This reverts commit bcd1a5b5cf2db8e86311d0c7fc513d1b84cf2714.

I'm reverting, instead of just deleting the commit locally, because the
code the last commit might be useful later.

I could also just leave this commit as single standalone branch, but
it might be hard to decide when to delete it.

So I'm just integrating these code to main's history without really
affecting the code by reverting.