Change 'Skin' and 'Skin_Full' schema
'price' and 'sale_price' in 'Skin' and 'Skin_Full' is being deprecated.
Instead, a 'last_price_history' is added.
It is either null indiciating there's not a single price history for
this item, or it is a 'Price History' schema type.

'price' and 'sale_price' will still be available for a while.