Rework run script
Completely reworked 'run.sh'.

'run.sh' now support two arguments
--prod: to launch app with production config
--port: to set listening port with 8000 as default

In production config, the script expects itself to be run with virtual
environment off.
It will start the app with nohup in the background, also logging the
output(both stdout and stderr) with timestamp in a file named
'uvicorn_[START_TIME].log'.
The log file is appended on each run.
Use log file name to easily locate the start of last run.

Original 'logs' directory and 'log_conf.json' is no longer used.

In non-production config, the script expects itself to be run with
virtual environment on, which is often activated while developing.
The output is not redirected or timestamped either.