Add initial support to read db config from Env Var
The app first try to read from 'db_config.ini' as it was.
If failed for any reason, the app try to read from environment variable.
The environment variable name is currently hardcoded, and it is expected
to be injected when used with docker.
Specifically, the deployment on AWS ECS use AWS SSM to manage these
environment variables.
We need a better approach to have a more centralized management of these
variables in later commits.